+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -133,6 +133,9 @@ Changes in existing checks
 Renamed checks
 ^^^^^^^^^^^^^^
 
+- The 'fuchsia-restrict-system-headers' check was renamed to :doc:`portability-restrict-system-includes
+  <clang-tidy/checks/portability-restrict-system-includes>
+
 Improvements to include-fixer
 -----------------------------